Crud cupboards and confirm message when deleting

+++ resources/views/pages/aliments/index.blade.php
@@ -1,7 +1,6 @@
 @extends('layouts.master')
 
 @section('content')
-
     <h1>All your food is here</h1>
     Do you wan't to add an aliment : <a href="{{ route('aliments.create') }}" class="btn btn-xs btn-primary">Add an aliment</a>
 
@@ -11,7 +10,7 @@
             <h2>
                 {!! Form::open(['method' => 'GET', 'url' => route('aliments.create', ['category_id' => $category->id])]) !!}
                     {!! Form::label('name', ucfirst($category->name)) !!}
-                    <button class="btn btn-xs btn-primary">+</button>
+                    <button class="btn btn-xs btn-primary">Add</button>
                 {!! Form::close() !!}
             </h2>
             @foreach ($aliments as $aliment)
@@ -31,8 +30,8 @@
                     <div class="col-sm-1">
                       <div class="col-sm-2"><a href="{{ route('aliments.edit', $aliment) }}" class="btn btn-xs btn-warning">Edit</a></div>
                       <div class="col-sm-2">
-                          {!! Form::open(['method' => 'DELETE', 'url' => route('aliments.destroy', $aliment->id)]) !!}
-                          <button class="btn btn-xs btn-danger">Delete</button>
+                          {!! Form::open(['method' => 'DELETE', 'url' => route('aliments.destroy', $aliment->id), 'onsubmit' => 'return ConfirmDelete()']) !!}
+                              <button class="btn btn-xs btn-danger">Delete</button>
                           {!! Form::close() !!}
                       </div>
                     </div>
